requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
# psycopg2-binary==2.9.9  # Только если используете PostgreSQL (раскомментируйте при необходимости)

